✨ refactor: rename llmkira sender kook files to app 🔧 Chore: update llmkira sender kook and slack tool list 🔒 security: update llmkira sender kook auth functionality
