+++ app/sender/kook/event.py
@@ -3,7 +3,7 @@
 # @Author  : sudoskys
 # @File    : event.py
 # @Software: PyCharm
-from llmkira.setting.discord import BotSetting
+from app.setting.discord import BotSetting
 
 _upload_error_message_template = [
     "I cant upload file {filename} to server {error}",
@@ -16,7 +16,7 @@ _upload_error_message_template = [
     "OMG, {filename} ,ERROR UPLOAD, `{error}`",
     "WTF, I CANT UPLOAD {filename} BECAUSE `{error}`",
     "MY PHONE IS BROKEN, I CANT UPLOAD {filename} BECAUSE `{error}`",
-    "As a human, I can't upload {filename} for you :( \n `{error}`"
+    "As a human, I can't upload {filename} for you :( \n `{error}`",
 ]
 
 
@@ -26,24 +26,18 @@ class MappingDefault(dict):
 
 
 def help_message():
-    return """ 
+    return """
     `{prefix}chat` - Chat with me :)
     `{prefix}task` - Ask me do things with `func_enable`
-    
+
     **Slash Command**
     `/help` - **You just did it :)**
     `/tool` - Check all useful tools
     `/clear` - wipe memory of your chat
     `/auth` - activate a task (my power)
-    `/bind` - bind third party platform
-    `/unbind` - unbind platform
-    `/set_endpoint` - <apikey>#<endpoint>
-    `/clear_endpoint` - clear endpoint and key
+    `/login` - login openai
+    `/login_via_url` - login via provider url
     `/env` - set environment variable
-    `/token` - bind your service provider token
-    `/token_clear` - clear your service provider token
-    `/func_ban` - ban function
-    `/func_unban` - unban function
 
 **Please confirm that that bot instance is secure, some plugins may be dangerous on unsafe instance.**
 """.format(prefix=BotSetting.prefix)
